move update logic to config/extensions/update.ts
